Refactoring space  model

+++ app/models/space.rb
@@ -23,40 +23,48 @@ class Space < ActiveRecord::Base
 
   before_save :name_format
 
+  ##
+  # add a new space
+  #
   def self.add_space(space_params, user_id, user_owner_id)
-    @space = Space.new(space_params)
-    @space.user_id =user_id
-    @space.user_owner_id = user_owner_id
+    space = Space.new(space_params)
+    space.user_id = user_id
+    space.user_owner_id = user_owner_id
 
-    @space.save!
+    space.save!
   end
 
+  ##
+  # edit a space
+  #
   def self.edit_space(space, space_params)
-    space.update(space_params)
+    space.update!(space_params)
   end
 
-  def self.delete_space(space, space_params)
-    not_found if space.name != space_params[:name]
-    # throw exception
-    not_found if Project.where(space_id: space.id).any?
+  ##
+  # delete a space
+  #
+  def self.delete_space(space, namespace)
+    raise StandardError, 'The space name is not valid' if space.name != namespace
+    if Project.has_projects? space.id
+      raise StandardError, 'You have to transfer or your projects or delete their'
+    end
 
     space.destroy!
   end
 
   ##
-  # Transfer space and projects to this team member
+  # Transfer space and projects to a member team
   #
   def self.transfer(space, user, user_member_id)
-    # TODO: if he is not my team member, throw exception
-    my_team = Team.where(user_id: user.id).where(user_team_id: user_member_id).any?
-    not_found unless my_team
+    raise StandardError, 'The member is not valid' unless Team.member? user.id, user_member_id
 
-    space.update(user_id: user_member_id)
-    Project.where(space_id: space.id).each do |p|
-      p.update_attribute(:user_id, user_member_id)
+    space.update!(user_id: user_member_id)
+    Project.find_each('space_id = ?', space.id) do |p|
+      p.update!(user_id: user_member_id)
     end
 
-    member_email = user_email(user_member_id)
+    member_email = Email.my_principal user_member_id
     Notifier.send_transfer_space_email(user, space, member_email).deliver_later unless member_email.nil?
   end
